[DEV] Updated interface of renderer. Makes it easier for app to specifiy draws
I want to have a renderer that will get a bunch of recorddraw functions
calls. When everything is recorded, a call of endframe will submit
command list to GPU for rendering. The renderer will not keep track of
objects that are on screen, that's the app's job. Communicate what is to
be drawn to Renderer layer

[NOTESskldjf